Updated everything to use denominations

+++ src/CalibrationOffsetHandler.cpp
@@ -2,7 +2,7 @@
 
 CalibrationOffsetHandler::CalibrationOffsetHandler(SharedData* sharedData) {
     this->sharedData = sharedData;
-    float offset = NVS.getFloat(KEY_OFFSET);
+    float offset = NVS.getInt(KEY_OFFSET);
     sharedData->setOffset(offset);
 }
 
@@ -12,8 +12,8 @@ CalibrationOffsetHandler::~CalibrationOffsetHandler() {
 void CalibrationOffsetHandler::tick() {
     if(sharedData->getState() == MachineState::SETTINGS_OFFSET_ADJUSTING &&
         sharedData->enterButton->pressed()) {
-            float oldOffset = NVS.getFloat(KEY_OFFSET);
-            double diff = sharedData->getOffset() - oldOffset;
+            float oldOffset = NVS.getInt(KEY_OFFSET);
+            long diff = sharedData->getOffset() - oldOffset;
             sharedData->setCurrentPosition(sharedData->getCurrentPosition() + diff);
             sharedData->setTargetPosition(sharedData->getCurrentPosition());
             NVS.setFloat(KEY_OFFSET, sharedData->getOffset());
